충돌

+++ jdbc_workshop1/jdbc_workshop1/src/main/java/com/nhnacademy/edu/jdbc1/service/course/Course.java
@@ -6,12 +6,12 @@ import java.util.Date;
 
 @Data
 public class Course {
-    private final int id;
-    private final int teacherId;
-    private final int subjectId;
+    private final Long id;
+    private final Long teacherId;
+    private final Long subjectId;
     private final Date createdAt;
 
-    public Course(int id, int teacherId, int subjectId, Date createdAt) {
+    public Course(Long id, Long teacherId, Long subjectId, Date createdAt) {
         this.id = id;
         this.teacherId = teacherId;
         this.subjectId = subjectId;
