+++ jdbc_workshop1/jdbc_workshop1/src/main/java/com/nhnacademy/edu/jdbc1/service/course/CourseRepository.java
@@ -4,13 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CourseRepository {
-    Optional<Course> findById(int id);
+    Optional<Course> findById(Long id);
 
     List<Course> findAll();
 
     int insert(Course user);
 
-    int delete(int id);
+    int delete(Long id);
 
-    int modify(int id, int teacherId,int subjectId);
+    int modify(Long id, Long teacherId,Long subjectId);
 }
